PO-2831 : TestHelperUtil (2)
