+++ src/integrationTest/java/uk/gov/hmcts/reform/opal/controllers/UserPermissionsV2LegacySyncIntegrationTest.java
@@ -12,7 +12,6 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.jdbc.Sql;
 import uk.gov.hmcts.reform.opal.AbstractIntegrationTest;
 import uk.gov.hmcts.reform.opal.LegacyWireMockXmlStubHelper;
-import uk.gov.hmcts.reform.opal.entity.BusinessEventLogType;
 import uk.gov.hmcts.reform.opal.entity.UserEntity;
 import uk.gov.hmcts.reform.opal.repository.UserRepository;
 import uk.gov.hmcts.reform.opal.service.synchronise.TestHelperService;
@@ -92,14 +91,13 @@ class UserPermissionsV2LegacySyncIntegrationTest extends AbstractIntegrationTest
 
         assertThat(roleCountBefore).isGreaterThan(0L);
 
-        mockMvc.perform(get(TestHelperUtil.userStateUri(user.getUserId())))
+        mockMvc.perform(get(userStateUri(user.getUserId())))
             .andExpect(status().isOk());
 
         testHelperService.assertUserRoleCount(user.getUserId(), 0L);
         testHelperService.assertUserHasNoActivationDate(user.getUserId());
         testHelperService.assertLoggedBusinessEventTypes();
     }
-
 
     @Test
     @DisplayName("AC2: should create missing BUU and assign role from cache when legacy returns BUU")
@@ -113,7 +111,7 @@ class UserPermissionsV2LegacySyncIntegrationTest extends AbstractIntegrationTest
 
         assertThat(testHelperService.businessUnitUserExists(BUSINESS_UNIT_USER_ID)).isFalse();
 
-        mockMvc.perform(get(TestHelperUtil.userStateUri(user.getUserId())));
+        mockMvc.perform(get(userStateUri(user.getUserId())));
 
         testHelperService.assertBusinessUnitUserRow(BUSINESS_UNIT_USER_ID, BUSINESS_UNIT_ID, USER_ID, ROLE_ID, 1L);
         testHelperService.assertLoggedBusinessEventTypes(ROLE_ASSIGNED_TO_USER, ACCOUNT_ACTIVATION_INITIATED);
@@ -131,7 +129,7 @@ class UserPermissionsV2LegacySyncIntegrationTest extends AbstractIntegrationTest
 
         testHelperService.assertBusinessUnitUserRow(EXISTING_BUSINESS_UNIT_USER_ID, LEGACY_BUSINESS_UNIT_ID, DIFFERENT_USER_ID);
 
-        mockMvc.perform(get(TestHelperUtil.userStateUri(user.getUserId())))
+        mockMvc.perform(get(userStateUri(user.getUserId())))
             .andExpect(status().isOk());
 
         testHelperService.assertBusinessUnitUserRow(EXISTING_BUSINESS_UNIT_USER_ID, LEGACY_BUSINESS_UNIT_ID, USER_ID, ROLE_ID, 1L);
@@ -151,7 +149,7 @@ class UserPermissionsV2LegacySyncIntegrationTest extends AbstractIntegrationTest
 
         testHelperService.assertBusinessUnitUserRow(EXISTING_BUSINESS_UNIT_USER_ID, LEGACY_BUSINESS_UNIT_ID, USER_ID);
 
-        mockMvc.perform(get(TestHelperUtil.userStateUri(user.getUserId())))
+        mockMvc.perform(get(userStateUri(user.getUserId())))
             .andExpect(status().isOk());
 
         testHelperService.assertBusinessUnitUserRow(EXISTING_BUSINESS_UNIT_USER_ID, BUSINESS_UNIT_ID, USER_ID, ROLE_ID, 1L);
@@ -173,7 +171,7 @@ class UserPermissionsV2LegacySyncIntegrationTest extends AbstractIntegrationTest
 
         assertThat(roleCountBefore).isGreaterThan(0L);
 
-        mockMvc.perform(get(TestHelperUtil.userStateUri(user.getUserId())))
+        mockMvc.perform(get(userStateUri(user.getUserId())))
             .andExpect(status().isOk());
 
         testHelperService.assertUserRoleCount(user.getUserId(), 0L);
@@ -206,7 +204,7 @@ class UserPermissionsV2LegacySyncIntegrationTest extends AbstractIntegrationTest
             ROLE_MAPPING_USER_PREFIX
         );
 
-        mockMvc.perform(get(TestHelperUtil.userStateUri(user.getUserId())))
+        mockMvc.perform(get(userStateUri(user.getUserId())))
             .andExpect(status().isOk());
 
         testHelperService.assertUserBusinessUnitIds(user.getUserId(), buId1, buId2);
@@ -243,7 +241,7 @@ class UserPermissionsV2LegacySyncIntegrationTest extends AbstractIntegrationTest
         testHelperService.setRoleMappingCache(user, roleMappingCache, ROLE_MAPPING_USER_PREFIX);
 
         //assertions on the response verify changes from the inner transaction are visible to outer calling code
-        mockMvc.perform(get(TestHelperUtil.userStateUri(user.getUserId())))
+        mockMvc.perform(get(userStateUri(user.getUserId())))
             .andExpect(status().isOk())
             .andExpect(jsonPath(
                 "$.domains.*.business_unit_users[*].business_unit_id",
@@ -286,7 +284,7 @@ class UserPermissionsV2LegacySyncIntegrationTest extends AbstractIntegrationTest
             ROLE_MAPPING_USER_PREFIX
         );
 
-        mockMvc.perform(get(TestHelperUtil.userStateUri(user.getUserId())))
+        mockMvc.perform(get(userStateUri(user.getUserId())))
             .andExpect(status().isOk());
 
         testHelperService.assertUserBusinessUnitIds(user.getUserId(), buId1, buId2, buId3);
@@ -323,7 +321,7 @@ class UserPermissionsV2LegacySyncIntegrationTest extends AbstractIntegrationTest
         Map<Long, Set<Short>> roleMappingCache = Map.of(ROLE_ID, Set.of(buId1, buId2, buId3));
         testHelperService.setRoleMappingCache(user, roleMappingCache, ROLE_MAPPING_USER_PREFIX);
 
-        mockMvc.perform(get(TestHelperUtil.userStateUri(user.getUserId())))
+        mockMvc.perform(get(userStateUri(user.getUserId())))
             .andExpect(status().isOk());
 
         testHelperService.assertUserBusinessUnitIds(user.getUserId(), buId1, buId2, buId3);
@@ -378,7 +376,7 @@ class UserPermissionsV2LegacySyncIntegrationTest extends AbstractIntegrationTest
             ROLE_MAPPING_USER_PREFIX
         );
 
-        mockMvc.perform(get(TestHelperUtil.userStateUri(user.getUserId())))
+        mockMvc.perform(get(userStateUri(user.getUserId())))
             .andExpect(status().isOk());
 
         testHelperService.assertUserBusinessUnitIds(user.getUserId(), buId1, buId2, buId3, buId4);
@@ -420,7 +418,7 @@ class UserPermissionsV2LegacySyncIntegrationTest extends AbstractIntegrationTest
         assertThat(testHelperService.businessUnitUserExists(BUSINESS_UNIT_USER_ID)).isFalse();
         long roleCountBefore = testHelperService.countRoleAssignments(user.getUserId());
 
-        mockMvc.perform(get(TestHelperUtil.userStateUri(user.getUserId())))
+        mockMvc.perform(get(userStateUri(user.getUserId())))
             .andExpect(status().isInternalServerError())
             .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_PROBLEM_JSON))
             .andExpect(jsonPath("$.title").value("Internal Server Error"))
@@ -436,4 +434,7 @@ class UserPermissionsV2LegacySyncIntegrationTest extends AbstractIntegrationTest
         testHelperService.assertLoggedBusinessEventTypes(ROLE_ASSIGNED_TO_USER);
     }
 
+    private String userStateUri(long userId) {
+        return "/v2/users/" + userId + "/state";
+    }
 }
